USB3Hub - Update commit - Hour 1
Started the schematic and some wiring. It may look like not much work was done, but I had to find from the parts library on the best USB 3.0 capable host controller for my needs as well as finding a suitable usb3.0 female port that support 3d view. Also connected the net portals and renamed them to unique names for the usb2Dp/m and usb3tx/rx/p/m